adds user online

+++ routes/api.php
@@ -17,37 +17,40 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::post('/group', 'ApiGroupsController@store');
-Route::patch('/group/{group}', 'ApiGroupsController@update');
-Route::get('/group/user/{user}', 'ApiGroupsController@index');
-Route::get('/group/{group}', 'ApiGroupsController@show');
-Route::get('/group/{group}/users', 'ApiGroupsController@indexusers');
-Route::get('/group/{group}/getdistrict', 'ApiGroupsController@getdistrict');
-Route::get('/group/{group}/removescan/{scan}', 'ApiGroupsController@removescan');
-Route::get('/grouprequest/{group}', 'ApiGrouprequestsController@index');
-Route::get('/grouprequest/{grouprequest}/accept', 'ApiGrouprequestsController@accept');
-Route::get('/grouprequest/{grouprequest}/deny', 'ApiGrouprequestsController@deny');
+Route::post('/group', 'ApiGroupsController@store')->middleware('auth:api');
+Route::patch('/group/{group}', 'ApiGroupsController@update')->middleware('auth:api');
+Route::get('/group/user/{user}', 'ApiGroupsController@index')->middleware('auth:api');
+Route::get('/group/{group}', 'ApiGroupsController@show')->middleware('auth:api');
+Route::get('/group/{group}/users', 'ApiGroupsController@indexusers')->middleware('auth:api');
+Route::post('/group/{group}/updatedistricts', 'ApiGroupsController@updatedistricts')->middleware('auth:api');
+Route::get('/group/{group}/removescan/{scan}', 'ApiGroupsController@removescan')->middleware('auth:api');
+Route::get('/grouprequest/{group}', 'ApiGrouprequestsController@index')->middleware('auth:api');
+Route::get('/grouprequest/{grouprequest}/accept', 'ApiGrouprequestsController@accept')->middleware('auth:api');
+Route::get('/grouprequest/{grouprequest}/deny', 'ApiGrouprequestsController@deny')->middleware('auth:api');
 
-Route::get('/dashmessage/user/{user}/', 'ApiDashmessagesController@index');
-Route::delete('/dashmessage/{dashmessage}', 'ApiDashmessagesController@delete');
+Route::get('/dashmessage/user/{user}/', 'ApiDashmessagesController@index')->middleware('auth:api');
+Route::delete('/dashmessage/{dashmessage}', 'ApiDashmessagesController@delete')->middleware('auth:api');
 
-Route::get('/scan/user/{user}', 'ApiScansController@indexuser');
-Route::get('/scan', 'ApiScansController@index');
-Route::get('/scan/{scan}', 'ApiScansController@show');
-Route::post('/scan/{scan}', 'ApiScansController@update');
-Route::post('/scan/', 'ApiScansController@store');
-Route::post('/scan/{scan}/withanswers', 'ApiScansController@storewithanswers');
-Route::get('/scan/{scan}/themes', 'ApiScansController@indexthemes');
-Route::get('/scan/{scan}/answers', 'ApiScansController@indexanswers');
-Route::post('/scan/{scan}/answers/{answer}', 'ApiScansController@postanswer');
-Route::post('/answer/{answer}', 'ApiAnswersController@store');
-Route::get('/scan/{scan}/theme/{theme}/questions', 'ApiScansController@indexquestions');
-Route::get('/scan/{scan}/user/', 'ApiScansController@getuser');
+Route::get('/scan/user/{user}', 'ApiScansController@indexuser')->middleware('auth:api');
+Route::get('/scan', 'ApiScansController@index')->middleware('auth:api');
+Route::post('/scan/', 'ApiScansController@store')->middleware('auth:api');
+Route::get('/scan/{scan}', 'ApiScansController@show')->middleware('auth:api');
+Route::post('/scan/{scan}', 'ApiScansController@update')->middleware('auth:api');
+Route::get('/scan/{scan}/getdistricts', 'ApiScansController@getdistricts')->middleware('auth:api');
+Route::post('/scan/{scan}/updatedistricts', 'ApiScansController@updatedistricts')->middleware('auth:api');
+Route::post('/scan/{scan}/withanswers', 'ApiScansController@storewithanswers')->middleware('auth:api');
+Route::get('/scan/{scan}/themes', 'ApiScansController@indexthemes')->middleware('auth:api');
+Route::get('/scan/{scan}/answers', 'ApiScansController@indexanswers')->middleware('auth:api');
+Route::post('/scan/{scan}/answers/{answer}', 'ApiScansController@postanswer')->middleware('auth:api');
+Route::post('/answer/{answer}', 'ApiAnswersController@store')->middleware('auth:api');
+Route::get('/scan/{scan}/theme/{theme}/questions', 'ApiScansController@indexquestions')->middleware('auth:api');
+Route::get('/scan/{scan}/user/', 'ApiScansController@getuser')->middleware('auth:api');
 
-Route::resource('measure', 'ApiMeasuresController');
-Route::resource('district', 'ApiDistrictsController');
-Route::resource('compares', 'ApiComparesController');
-Route::get('/compare/{compare}/scan/{scan}', 'ApiComparesController@destroycompare');
-Route::post('/measure/{measure}/user/{user}', 'ApiMeasuresController@adduser');
-Route::get('/measure/{measure}/user/{user}/removeuser', 'ApiMeasuresController@removeuser');
+Route::resource('measure', 'ApiMeasuresController')->middleware('auth:api');
+Route::resource('district', 'ApiDistrictsController')->middleware('auth:api');
+Route::resource('compares', 'ApiComparesController')->middleware('auth:api');
+
+Route::get('/compare/{compare}/scan/{scan}', 'ApiComparesController@destroycompare')->middleware('auth:api');
+Route::post('/measure/{measure}/user/{user}', 'ApiMeasuresController@adduser')->middleware('auth:api');
+Route::get('/measure/{measure}/user/{user}/removeuser', 'ApiMeasuresController@removeuser')->middleware('auth:api');
